Add support to change default-branch for github projects
This is the first step on gitops for branch rename in github. Most
projects no longer wish to use 'master' as the default branch name. As
such, we need a way to update it.

Currently, there is no way to auto_init a repo in github using a branch
other then master.  We'll have to work around this.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>