Revert ansible-changelog changes
This is clearly broken, and no longer checks to see if we are adding in
changelog fragments.

Given we don't have the time to properly debug this, we are going to
revert.

That said, the proper way to do this moving forward, is to create a new
job, which the logic that is needed. Then smoke test the heck out of it
on a specific repo.

Then, over time we can replace all projects with it.  Sadly, this was
rushed into service before. Not laying fault, but think we need to do a
better job at catching all new features next time.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>