use gh to compare Github branch configuration
Use the `gh` CLI tool to query Github API to get the right default
branch. Use this information to compare wit the local configuration.

Also properly compare the name of the repository,
ansible-community/molecule is in
github.com/ansible-community/molecule-libvirt but are not the same.

Finally use an emoji to mark the sections instead of a series of '='. Those
where getting too long.